fix: Support INTERVAL expressions in SQLite
